+++ R/LassoLogisticRegression.R
@@ -20,13 +20,16 @@
 #'
 #' @param variance   a single value used as the starting value for the automatic lambda search
 #' @param seed       An option to add a seed when training the model
+#' @param threads    An option to set number of threads when training model
 #'
 #' @examples
 #' model.lr <- setLassoLogisticRegression()
 #' @export
-setLassoLogisticRegression<- function(variance=0.01, seed=NULL){
+setLassoLogisticRegression<- function(variance=0.01, seed=NULL, threads = -1){
   if(!class(seed)%in%c('numeric','NULL','integer'))
     stop('Invalid seed')
+  if(!class(threads)%in%c('numeric','NULL','integer'))
+    stop('Invalid threads')
   if(!class(variance) %in% c("numeric", "integer"))
     stop('Variance must be numeric')
   if(variance<0)
@@ -37,7 +40,7 @@ setLassoLogisticRegression<- function(variance=0.01, seed=NULL){
     seed <- as.integer(sample(100000000,1))
   }
   
-  result <- list(model='fitLassoLogisticRegression', param=list(variance=variance, seed=seed[1]), name="Lasso Logistic Regression")
+  result <- list(model='fitLassoLogisticRegression', param=list(variance=variance, seed=seed[1], threads = threads[1]), name="Lasso Logistic Regression")
   class(result) <- 'modelSettings' 
   
   return(result)
@@ -76,7 +79,7 @@ fitLassoLogisticRegression<- function(population, plpData, param, search='adapti
                                                              tolerance  = 2e-07,
                                                              cvRepetitions = 1, fold=ifelse(!is.null(population$indexes),max(population$indexes),1),
                                                              selectorType = "byPid",
-                                                             threads=-1,
+                                                             threads= param$threads,
                                                              maxIterations = 3000,
                                                              seed=param$seed))
   
@@ -104,12 +107,19 @@ fitLassoLogisticRegression<- function(population, plpData, param, search='adapti
               population = population, 
               plpData = plpData)
   
+  # get cv AUC
+  cvPred <- do.call(rbind, lapply(modelTrained$cv, function(x){x$predCV}))
+  cvPerFold <-  unlist(lapply(modelTrained$cv, function(x){x$out_sample_auc}))
+  names(cvPerFold) <- paste0('fold_auc', 1:length(cvPerFold))
+  
   result <- list(model = modelTrained,
                  modelSettings = list(model='lr_lasso', modelParameters=param), #todo get lambda as param
                  hyperParamSearch = c(priorVariance=modelTrained$priorVariance, 
                                       seed=ifelse(is.null(param$seed), 'NULL', param$seed  ), 
-                                      log_likelihood = modelTrained$log_likelihood),
-                 trainCVAuc = NULL,
+                                      log_likelihood = modelTrained$log_likelihood,
+                                      cvPerFold,
+                                      auc = aucWithoutCi(cvPred$value, cvPred$y)),
+                 trainCVAuc = cvPerFold,
                  metaData = plpData$metaData,
                  populationSettings = attr(population, 'metaData'),
                  outcomeId=outcomeId,# can use populationSettings$outcomeId?
@@ -124,3 +134,5 @@ fitLassoLogisticRegression<- function(population, plpData, param, search='adapti
   attr(result, 'predictionType') <- 'binary'
   return(result)
 }
+
+